Add test for circular dependencies Fix circular dependencies
