+++ package.json
@@ -74,6 +74,7 @@
     "copyfiles": "^2.4.1",
     "eslint": "7.32.0",
     "eslint-plugin-oclif": "0.1.0",
+    "madge": "^5.0.1",
     "mocha": "^9.1.2",
     "mock-fs": "^5.1.1",
     "nock": "^13.1.3",
@@ -124,6 +125,7 @@
   "scripts": {
     "lint": "eslint --ext .ts,.js .",
     "fix": "eslint --ext .ts,.js . --fix",
+    "check-circular": "madge --ts-config ./tsconfig.json --circular --extensions ts ./src/",
     "build": "rm -rf ./lib && rm -f tsconfig.tsbuildinfo && tsc -b && npm run copy-static",
     "copy-static": "copyfiles -u 1 \"src/static/*.html\" lib",
     "postpack": "rm -f oclif.manifest.json",
