+++ src/app-config/auth.ts
@@ -1,8 +1,8 @@
 import { AccessToken, AuthorizationCode } from 'simple-oauth2';
 import { URL } from 'url';
+import User from '../architect/user/user.entity';
 import LoginRequiredError from '../common/errors/login-required';
 import { docker } from '../common/utils/docker';
-import { User } from '../common/utils/user';
 import CallbackServer from './callback_server';
 import AppConfig from './config';
 import CredentialManager from './credentials';
