+++ src/app-config/service.ts
@@ -3,8 +3,8 @@ import fs from 'fs-extra';
 import https from 'https';
 import path from 'path';
 import { URL } from 'url';
+import User from '../architect/user/user.entity';
 import LoginRequiredError from '../common/errors/login-required';
-import { User } from '../common/utils/user';
 import { Dictionary } from '../dependency-manager/src/utils/dictionary';
 import LocalPaths from '../paths';
 import AuthClient from './auth';
