+++ src/architect/account/account.entity.ts
@@ -1,0 +1,4 @@
+export default interface Account {
+  id: string;
+  name: string;
+}
