+++ src/architect/deployment/deployment.entity.ts
@@ -1,6 +1,4 @@
-import { Pipeline } from "./pipeline";
-
-export interface Deployment {
+export default interface Deployment {
   id: string;
   instance_id: string;
   applied_at?: string;
