+++ src/architect/environment/environment.entity.ts
@@ -1,0 +1,7 @@
+import Platform from '../platform/platform.entity';
+
+export default interface Environment {
+  id: string;
+  name: string;
+  platform: Platform;
+}
