+++ src/architect/platform/platform.entity.ts
@@ -1,0 +1,6 @@
+export default interface Platform {
+  id: string;
+  name: string;
+  type: string;
+  account: Account;
+}
