+++ src/architect/user/user.entity.ts
@@ -1,4 +1,4 @@
-export interface User {
+export default interface User {
   id: string;
   name: string;
 }
