+++ src/commands/deploy.ts
@@ -8,15 +8,15 @@ import inquirer from 'inquirer';
 import isCi from 'is-ci';
 import yaml from 'js-yaml';
 import opener from 'opener';
+import AccountUtils from '../architect/account/account.utils';
+import { EnvironmentUtils } from '../architect/environment/environment.utils';
+import PipelineUtils from '../architect/pipeline/pipeline.utils';
 import Command from '../base-command';
 import LocalDependencyManager from '../common/dependency-manager/local-manager';
 import { DockerComposeUtils } from '../common/docker-compose';
 import DockerComposeTemplate from '../common/docker-compose/template';
 import { DeploymentFailedError, PipelineAbortedError, PollingTimeout } from '../common/errors/pipeline-errors';
-import { AccountUtils } from '../common/utils/account';
 import * as Docker from '../common/utils/docker';
-import { EnvironmentUtils } from '../common/utils/environment';
-import { PipelineUtils } from '../common/utils/pipeline';
 import { ComponentSlugUtils, ComponentSpec, ComponentVersionSlugUtils } from '../dependency-manager/src';
 import { buildSpecFromPath } from '../dependency-manager/src/spec/utils/component-builder';
 import { Dictionary } from '../dependency-manager/src/utils/dictionary';
