+++ src/commands/register.ts
@@ -8,9 +8,9 @@ import isCi from 'is-ci';
 import path from 'path';
 import tmp from 'tmp';
 import untildify from 'untildify';
+import AccountUtils from '../architect/account/account.utils';
 import Command from '../base-command';
 import MissingContextError from '../common/errors/missing-build-context';
-import { AccountUtils } from '../common/utils/account';
 import * as Docker from '../common/utils/docker';
 import { oras } from '../common/utils/oras';
 import { ArchitectError, ComponentSlugUtils, Refs, ResourceSpec, Slugs } from '../dependency-manager/src';
