+++ src/common/dependency-manager/local-manager.ts
@@ -2,8 +2,7 @@ import { AxiosInstance } from 'axios';
 import chalk from 'chalk';
 import deepmerge from 'deepmerge';
 import yaml from 'js-yaml';
-import DependencyManager, { buildSpecFromYml, ComponentSlugUtils, ComponentSpec, ComponentVersionSlugUtils, IngressSpec } from '../../dependency-manager/src';
-import { ComponentInstanceMetadata } from '../../dependency-manager/src/config/component-config';
+import DependencyManager, { buildSpecFromYml, ComponentInstanceMetadata, ComponentSlugUtils, ComponentSpec, ComponentVersionSlugUtils, IngressSpec } from '../../dependency-manager/src';
 import DependencyGraph from '../../dependency-manager/src/graph';
 import { buildSpecFromPath } from '../../dependency-manager/src/spec/utils/component-builder';
 import { generateIngressesOverrideSpec, overrideSpec } from '../../dependency-manager/src/spec/utils/spec-merge';
