+++ src/common/utils/kubernetes-platform.utils.ts
@@ -6,7 +6,7 @@ import inquirer from 'inquirer';
 import yaml from 'js-yaml';
 import path from 'path';
 import untildify from 'untildify';
-import { CreatePlatformInput } from './platform';
+import { CreatePlatformInput } from '../../architect/platform/platform.utils';
 
 const SERVICE_ACCOUNT_NAME = 'architect';
