+++ src/dependency-manager/src/spec/utils/spec-merge.ts
@@ -2,9 +2,9 @@ import { plainToClass } from 'class-transformer';
 import deepmerge, { Options } from 'deepmerge';
 import { Dictionary } from '../../utils/dictionary';
 import { ValidationError, ValidationErrors } from '../../utils/errors';
+import { findPotentialMatch } from '../../utils/match';
 import { RecursivePartial } from '../../utils/types';
 import { ComponentSpec, IngressSpec } from '../component-spec';
-import { findPotentialMatch } from './spec-validator';
 
 export function generateIngressesOverrideSpec(component_spec: ComponentSpec, ingresses: Dictionary<IngressSpec>): RecursivePartial<ComponentSpec> {
   const spec: RecursivePartial<ComponentSpec> = {};
