+++ src/dependency-manager/src/spec/utils/spec-validator.ts
@@ -3,84 +3,15 @@ import ajv_errors from "ajv-errors";
 import addFormats from "ajv-formats";
 import { plainToClass } from 'class-transformer';
 import cron from 'cron-validate';
-import leven from 'leven';
 import { Dictionary } from '../../utils/dictionary';
 import { ValidationError, ValidationErrors } from '../../utils/errors';
 import { buildContextMap, replaceBrackets } from '../../utils/interpolation';
+import { findPotentialMatch } from '../../utils/match';
+import { ParsedYaml } from '../../utils/types';
 import { ComponentSpec } from '../component-spec';
-import { ParsedYaml } from './component-builder';
 import { findDefinition, getArchitectJSONSchema } from './json-schema';
 
 export type AjvError = ErrorObject[] | null | undefined;
-
-export const findPotentialMatch = (value: string, options: string[], max_distance = 15): string | undefined => {
-  let potential_match;
-  let shortest_distance = Infinity;
-  const value_length = value.length;
-  for (const option of [...options].sort()) {
-    const option_length = option.length;
-    // https://github.com/sindresorhus/leven/issues/14
-    if (Math.abs(value_length - option_length) >= max_distance) {
-      continue;
-    }
-
-    const distance = leven(value, option);
-    if (distance < max_distance && distance <= shortest_distance) {
-      potential_match = option;
-      shortest_distance = distance;
-    }
-  }
-  return potential_match;
-};
-
-function escapeRegex(string: string) {
-  return string.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&');
-}
-
-export const addLineNumbers = (value: string, errors: ValidationError[]): void => {
-  const rows = value.split('\n');
-  const total_rows = rows.length;
-  for (const error of errors) {
-    const keys = error.path.split('.');
-    let pattern = '(.*?)' + keys.map((key) => `${escapeRegex(key)}:`).join('(.*?)');
-
-    const target_value = `${error.value}`.split('\n')[0];
-    if (!error.invalid_key) {
-      pattern += `(.*?)${escapeRegex(target_value)}`;
-    }
-
-    const exp = new RegExp(pattern, 's');
-    const matches = exp.exec(value);
-    if (matches) {
-      const match = matches[0];
-      const remaining_rows = value.replace(match, '').split('\n');
-      const target_row = total_rows - remaining_rows.length;
-      const end_row = rows[target_row];
-
-      const end_length = (remaining_rows[0]?.length || 0);
-
-      if (error.invalid_key) {
-        error.start = {
-          row: target_row + 1,
-          column: (end_row.length - end_row.trimLeft().length) + 1,
-        };
-        error.end = {
-          row: target_row + 1,
-          column: end_row.length - end_length,
-        };
-      } else {
-        error.start = {
-          row: target_row + 1,
-          column: (end_row.length - (target_value.length + (end_length ? end_length - 1 : 0))),
-        };
-        error.end = {
-          row: target_row + 1,
-          column: end_row.length - end_length,
-        };
-      }
-    }
-  }
-};
 
 export const mapAjvErrors = (parsed_yml: ParsedYaml, ajv_errors: AjvError): ValidationError[] => {
   if (!ajv_errors?.length) {
