+++ src/dependency-manager/src/utils/regex.ts
@@ -1,0 +1,13 @@
+export const matches = (text: string, pattern: RegExp): { [Symbol.iterator]: () => Generator<RegExpExecArray, void, unknown>; } => ({
+  [Symbol.iterator]: function* () {
+    const clone = new RegExp(pattern.source, pattern.flags);
+    let match = null;
+    do {
+      match = clone.exec(text);
+      if (match) {
+        yield match;
+        clone.lastIndex = match.index + 1; // Support overlapping match groups
+      }
+    } while (match);
+  },
+});
